BookProject/requirements.txt
backports.csv==1.0.7
defusedxml==0.5.0
diff-match-patch==20181111
Django==1.11
django-crispy-forms==1.7.2
django-crontab==0.7.1
django-formtools==2.1
django-import-export==1.2.0
django-reversion==3.0.3
django-smtp-ssl==1.0
djangorestframework==3.9.2
et-xmlfile==1.0.1
future==0.17.1
graphic-verification-code==1.0.2
httplib2==0.9.2
jdcal==1.4
jieba==0.39
odfpy==1.4.0
openpyxl==2.6.2
Pillow==6.0.0
PyMySQL==0.9.3
pytz==2019.1
PyYAML==5.1
six==1.12.0
tablib==0.13.0
xadmin==0.6.1
xlrd==1.2.0
xlwt==1.3.0
